User permissions

+++ app/Http/Controllers/PostController.php
@@ -5,9 +5,12 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
 use App\Models\Post;
+use Illuminate\Auth\Access\Events\GateEvaluated;
+use Illuminate\Auth\Access\Gate as AccessGate;
 use Illuminate\Http\Request;
 use Illuminate\Routing\Controllers\HasMiddleware;
 use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller implements HasMiddleware
 {
@@ -52,6 +55,8 @@ class PostController extends Controller implements HasMiddleware
   
     public function update(Request $request, Post $post)
     {
+        
+        Gate::authorize('modify', $post);
         $validate = $request->validate([
             'title' => 'required|max:255',
             'body' => 'required'
@@ -66,6 +71,7 @@ class PostController extends Controller implements HasMiddleware
   
     public function destroy(Post $post)
     {
+        Gate::authorize('modify', $post);
         $post->delete($post);
         return response()->json([
             'message' => 'Post has deleted',
